+++ tests/conftest.py
@@ -52,7 +52,7 @@ async def mock_call_hockeytech_api(hass):
                 "url": url,
             }
 
-    with patch("custom_components.teamtracker.hockeytech.HockeyTechProvider.async_call_hockeytech_api", new_callable=AsyncMock) as mock_hockeytech:
+    with patch("custom_components.teamtracker.provide_hockeytech.HockeyTechProvider.async_call_hockeytech_api", new_callable=AsyncMock) as mock_hockeytech:
         mock_hockeytech.side_effect = _get_mock_ht_api_data
         yield mock_hockeytech
 
@@ -97,8 +97,8 @@ async def mock_call_espn_api(hass):
             return {"data": None, "url": url}
 
     # Patch the actual utility function
-    with patch("custom_components.teamtracker.espn.EspnProvider.async_call_espn_api", new_callable=AsyncMock) as mock_espn, \
-        patch("custom_components.teamtracker.espn_all_leagues.EspnAllLeaguesProvider.async_call_espn_api", new_callable=AsyncMock) as mock_ea:
+    with patch("custom_components.teamtracker.provide_espn.EspnProvider.async_call_espn_api", new_callable=AsyncMock) as mock_espn, \
+        patch("custom_components.teamtracker.provide_espn_all.EspnAllLeaguesProvider.async_call_espn_api", new_callable=AsyncMock) as mock_ea:
         mock_espn.side_effect = _get_mock_api_data
         mock_ea.side_effect = _get_mock_api_data
         yield mock_espn
