+++ tests/test_event_old.py
@@ -9,7 +9,7 @@ from custom_components.teamtracker.const import (
     DEFAULT_LAST_UPDATE,
     DEFAULT_LOGO,
 )
-from custom_components.teamtracker.event import EspnParser
+from custom_components.teamtracker.parse_espn import EspnParser
 
 from tests.const import TEST_DATA
